creates a new spacereport everytime

+++ pages/index.vue
@@ -10,25 +10,22 @@
         <section id="section_layout" class="grid grid-cols-2">
           <div class="m-2">
             <p>space reports created by {{ user.first_name }}</p>
-            <div id="report_container" class="border border-gray-900 h-60">
-              <div
-                v-for="(report, index) in reports"
-                :key="report"
-                class="secondary-button mx-3 bg-blue-300 rounded"
-              >
-                {{ reports.missionname }}
-                <button
-                  class="text-white bg-blue-800 px-2"
-                  @click="updateReport(index)"
+            <div id="report_container" class="border border-gray-900 h-full">
+              <p v-if="!reports.length">no spacereports saved</p>
+              <div v-if="reports.length">
+                <div
+                  v-for="(report, index) in reports"
+                  :key="report.missionname"
+                  class="secondary-button mx-3 bg-blue-300 rounded"
                 >
-                  update
-                </button>
-                <button
-                  class="text-white bg-black px-2"
-                  @click="removeReport(index)"
-                >
-                  x
-                </button>
+                  {{ report.missionname }}
+                  <button class="primary-button" @click="updateReport(report)">
+                    edit
+                  </button>
+                  <button class="secondary-button" @click="removeReport(index)">
+                    delete
+                  </button>
+                </div>
               </div>
             </div>
           </div>
@@ -58,34 +55,11 @@ export default {
 
   data() {
     return {
-      // thespacereports: [],
       user: '',
       reports: [],
     }
   },
-  // created() {
-  //   this.thespacereports = [
-  //     {
-  //       // fake space reports
-  //       id: 1,
-  //       missionname: 'first launch',
-  //       missiondesc: 'it went bad',
-  //       missiondate: '11-11-11',
-  //       selectedimages: 'pictures',
-  //       lat: '1234',
-  //       long: '1234',
-  //     },
-  //     {
-  //       id: 2,
-  //       missionname: 'second launch',
-  //       missiondesc: 'it went really bad bad',
-  //       missiondate: '11-11-11',
-  //       selectedimages: 'pictures',
-  //       lat: '1234',
-  //       long: '1234',
-  //     },
-  //   ]
-  // },
+
   mounted() {
     if (localStorage.user) {
       this.user = JSON.parse(localStorage.user)
@@ -94,27 +68,23 @@ export default {
     }
     if (localStorage.reports) {
       this.reports = JSON.parse(localStorage.reports)
-      console.log(this.reports.value)
+      console.log(this.reports[0])
     }
   },
   methods: {
-    // storage() {
-    //   if (process.browser) {
-    //     localStorage.getItem('authToken')
-    //   }
-    // },
     clearUser() {
       localStorage.clear()
       this.$router.push({ path: '/login' })
     },
     removeReport(index) {
       console.log('delete test')
-      console.log(this.reports)
-      this.reports.splice(index, 1)
-      localStorage.setItem('reports', JSON.stringify(this.reports))
-
-      // use splice here
-      // this.spacereports.splice(this.spacereports.indexOf(this.spacereports), 1)
+      let response = confirm(
+        `Are you sure you want to delete ${this.reports[0]}`
+      )
+      if (response) {
+        this.reports.splice(index, 1)
+        localStorage.setItem('reports', JSON.stringify(this.reports))
+      }
     },
     updateReport(index) {
       console.log('edit test')
